Move MarketplaceRow to its own file
